refactor: Define error in modules
