+++ src/lib.rs
@@ -32,6 +32,5 @@
 //! ```
 
 pub mod client;
-mod error;
 pub mod record;
 mod worker;
